tinggal membuat laporan dan hak akses

+++ src/views/Pegawai/MainPegawaiView.vue
@@ -115,6 +115,14 @@ const handleDelete = async (id) => {
   // Jika pengguna mengonfirmasi
   if (result.isConfirmed) {
     try {
+      const isInUse = await checkIfDataInUse(id)
+
+      // return console.log({ isInUse })
+      if (isInUse?.status) {
+        // Jika data digunakan di object store lain, tampilkan notifikasi error
+        Swal.fire({ icon: 'error', title: `Data gagal dihapus`, text: isInUse.message })
+        return
+      }
       // Menghapus item dari IndexedDB
       const deleteItem = await idbStore.deleteItemById(objectStore, id)
 
@@ -159,6 +167,33 @@ const modalProps = ref({
   actionType: 'add',
   data: {}
 })
+
+const checkIfDataInUse = async (id) => {
+  try {
+    // Misal ada beberapa object store lain seperti 'otherStore1' dan 'otherStore2'
+    const otherStores = ['delegasiPegawai']
+
+    for (const store of otherStores) {
+      const data = await idbStore.fetchData(store) // Mendapatkan semua data dari object store
+
+      // Periksa apakah data di dalam object store berbentuk array of objects
+
+      console.log({ dataarray: data, id })
+      // Jika array, periksa apakah ada object di dalam array yang memiliki ID yang sama
+      const isUsed = data.some((item) =>
+        item.relation?.some((relationItem) => relationItem.pegawai.id === id)
+      )
+      if (isUsed)
+        return { status: true, message: 'Data sedang digunakan pada menu Delegasi Pegawai' } // Jika ditemukan, data sedang digunakan
+    }
+
+    // Jika tidak ditemukan di object store lain, berarti data tidak digunakan
+    return { status: false }
+  } catch (error) {
+    console.error('Error checking if data is in use:', error)
+    return { status: false } // Jika terjadi error, anggap data tidak sedang digunakan
+  }
+}
 </script>
 
 <template>
